antigravity_system_IA_agents/main.py
"""
╔══════════════════════════════════════════════════════════════════════════════╗
║                    ANTIGRAVITY - SISTEMA MULTI-AGENTE IA                    ║
║                        Versión: 2.0 (2026 - 2050)                           ║
║                    Arquitectura Profesional de Largo Plazo                   ║
╚══════════════════════════════════════════════════════════════════════════════╝

Sistema de 4 agentes especializados trabajando en armonía:
1. AGENTE FRONTEND/BACKEND - Arquitectura técnica profesional
2. AGENTE SEGURIDAD - Cifrado, autenticación, protección
3. AGENTE SOFTWARE/DESIGN - UI/UX, Diseño responsivo, Estética
4. AGENTE ERRORES - Testing, Debugging, Monitoreo proactivo

Pensado para: 20+ años (2026-2050) sin obsolescencia
"""

import os
import sys
import json
import logging
from pathlib import Path
from datetime import datetime
from dotenv import load_dotenv
from typing import Dict, List, Any

# Cargar variables de entorno
load_dotenv()

# Configurar logging profesional
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('antigravity.log'),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

# Importar agentes
from agents.agent_frontend_backend import AgenteFrontendBackend
from agents.agent_security import AgenteSeguridad
from agents.agent_software_design import AgenteSoftwareDesign
from agents.agent_error_handler import AgenteErrorHandler
from agents.orchestrator import Orchestrator
from config.system_config import SYSTEM_CONFIG, PROJECT_TEMPLATES
from utils.validators import ValidarProyecto
from utils.report_generator import GenerarReporte


class AntugravitySystem:
    """Sistema central que orquesta todos los agentes"""
    
    def __init__(self):
        """Inicializar el sistema con todos los componentes"""
        logger.info("=" * 80)
        logger.info("INICIALIZANDO ANTIGRAVITY - SISTEMA MULTI-AGENTE")
        logger.info(f"Timestamp: {datetime.now().isoformat()}")
        logger.info("=" * 80)
        
        # Inicializar agentes
        self.agente_dev = AgenteFrontendBackend()
        self.agente_seguridad = AgenteSeguridad()
        self.agente_design = AgenteSoftwareDesign()
        self.agente_errores = AgenteErrorHandler()
        
        # Inicializar orquestador
        self.orchestrator = Orchestrator(
            agentes={
                'frontend_backend': self.agente_dev,
                'seguridad': self.agente_seguridad,
                'software_design': self.agente_design,
                'error_handler': self.agente_errores
            }
        )
        
        # Validador
        self.validador = ValidarProyecto()
        
        # Generador de reportes
        self.generador_reportes = GenerarReporte()
        
        logger.info("✓ Sistema inicializado correctamente")
        
    def crear_proyecto(self, especificaciones: Dict[str, Any]) -> Dict[str, Any]:
        """
        Crear un nuevo proyecto con todos los agentes
        
        Args:
            especificaciones: Diccionario con requisitos del proyecto
            
        Returns:
            Resultado completo del proyecto
        """
        logger.info("=" * 80)
        logger.info(f"CREANDO PROYECTO: {especificaciones.get('nombre', 'Sin nombre')}")
        logger.info("=" * 80)
        
        # Validar entrada
        validacion = self.validador.validar(especificaciones)
        if not validacion['valido']:
            logger.error(f"Errores de validación: {validacion['errores']}")
            return {'exito': False, 'errores': validacion['errores']}
        
        # Ejecutar flujo completo
        try:
            resultado = self.orchestrator.ejecutar_flujo_completo(especificaciones)
            
            # Generar reporte
            reporte = self.generador_reportes.generar(
                especificaciones=especificaciones,
                resultado=resultado
            )
            
            logger.info("✓ Proyecto creado exitosamente")
            return {
                'exito': True,
                'proyecto': resultado,
                'reporte': reporte
            }
            
        except Exception as e:
            logger.error(f"Error en creación de proyecto: {str(e)}", exc_info=True)
            return {'exito': False, 'error': str(e)}
    
    def analizar_codigo_existente(self, ruta_codigo: str) -> Dict[str, Any]:
        """
        Analizar código existente y mejorar con agentes
        
        Args:
            ruta_codigo: Ruta al código a analizar
            
        Returns:
            Análisis y mejoras propuestas
        """
        logger.info(f"Analizando código existente: {ruta_codigo}")
        
        try:
            # Leer código
            with open(ruta_codigo, 'r', encoding='utf-8') as f:
                codigo = f.read()
            
            # Analizar con cada agente
            analisis = {
                'frontend_backend': self.agente_dev.analizar_codigo(codigo),
                'seguridad': self.agente_seguridad.analizar_seguridad(codigo),
                'design': self.agente_design.analizar_diseño(codigo),
                'errores': self.agente_errores.analizar_errores(codigo)
            }
            
            logger.info("✓ Análisis completado")
            return analisis
            
        except Exception as e:
            logger.error(f"Error en análisis: {str(e)}")
            return {'error': str(e)}
    
    def obtener_estado_sistema(self) -> Dict[str, Any]:
        """Obtener estado actual de todos los agentes"""
        return {
            'timestamp': datetime.now().isoformat(),
            'agentes': {
                'frontend_backend': self.agente_dev.obtener_estado(),
                'seguridad': self.agente_seguridad.obtener_estado(),
                'software_design': self.agente_design.obtener_estado(),
                'error_handler': self.agente_errores.obtener_estado()
            },
            'sistema': SYSTEM_CONFIG
        }


def main():
    """Función principal"""
    print("\n" + "=" * 80)
    print("ANTIGRAVITY - SISTEMA PROFESIONAL DE AGENTES IA")
    print("Versión: 2.0 (2026-2050)")
    print("=" * 80 + "\n")
    
    # Inicializar sistema
    sistema = AntugravitySystem()
    
    # Ejemplo de proyecto
    proyecto_ejemplo = {
        'nombre': 'Dashboard Premium Pro',
        'descripcion': 'Dashboard administrativo profesional',
        'tecnologia': {
            'frontend': 'React 19 + TypeScript + Tailwind CSS',
            'backend': 'FastAPI + PostgreSQL',
            'deployment': 'Docker + Kubernetes'
        },
        'requisitos': {
            'responsivo': True,
            'offline_capable': True,
            'pwa': True,
            'dark_mode': True,
            'accesibilidad': 'WCAG 2.1 AAA'
        },
        'seguridad': {
            'autenticacion': 'OAuth 2.0 + JWT',
            'encryption': 'AES-256',
            'rate_limiting': True,
            'owasp_compliance': True
        },
        'dispositivos': ['mobile', 'tablet', 'desktop'],
        'performance': {
            'lighthouse_score': 95,
            'first_contentful_paint': 1200,
            'time_to_interactive': 2500
        }
    }
    
    # Crear proyecto
    resultado = sistema.crear_proyecto(proyecto_ejemplo)
    
    if resultado['exito']:
        print("\n✓ PROYECTO CREADO EXITOSAMENTE\n")
        print(json.dumps(resultado, indent=2, ensure_ascii=False))
    else:
        print(f"\n✗ Error: {resultado.get('error', 'Desconocido')}\n")
    
    # Mostrar estado del sistema
    print("\n" + "=" * 80)
    print("ESTADO DEL SISTEMA")
    print("=" * 80)
    estado = sistema.obtener_estado_sistema()
    print(json.dumps(estado, indent=2, ensure_ascii=False))


if __name__ == "__main__":
    main()
