antigravity_system_IA_agents/requirements.txt
# ANTIGRAVITY SYSTEM DEPENDENCIES
# Python 3.11+
# Instalar con: pip install -r requirements.txt

# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi==0.104.1           # Backend framework
uvicorn==0.24.0            # ASGI server
pydantic==2.5.0             # Data validation
python-dotenv==1.0.0       # Environment variables

# ============================================================================
# DATABASE & CACHING
# ============================================================================
sqlalchemy==2.0.23          # ORM
alembic==1.13.0            # Database migrations
psycopg2-binary==2.9.9     # PostgreSQL adapter
redis==5.0.1               # Redis client
motor==3.3.2               # Async MongoDB driver

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
python-jose[cryptography]==3.3.0  # JWT handling
cryptography==41.0.7       # Encryption
bcrypt==4.1.1              # Password hashing
PyJWT==2.8.1               # JWT tokens
oauthlib==3.2.2            # OAuth 2.0
authlib==1.3.0             # Authentication library
passlib[bcrypt]==1.7.4     # Password utilities

# ============================================================================
# TESTING & QUALITY ASSURANCE
# ============================================================================
pytest==7.4.3              # Testing framework
pytest-cov==4.1.0          # Coverage
pytest-asyncio==0.21.1     # Async testing
httpx==0.25.2              # HTTP client for testing
locust==2.17.0             # Load testing
black==23.12.0             # Code formatter
flake8==6.1.0              # Linter
mypy==1.7.1                # Type checker
pylint==3.0.3              # Code analysis

# ============================================================================
# MONITORING & LOGGING
# ============================================================================
sentry-sdk==1.38.0         # Error tracking
python-json-logger==2.0.7  # JSON logging
loguru==0.7.2              # Better logging
prometheus-client==0.19.0  # Prometheus metrics
opentelemetry-api==1.21.0  # Observability
opentelemetry-sdk==1.21.0

# ============================================================================
# API DOCUMENTATION
# ============================================================================
swagger-ui==1.7.3          # Swagger UI
mkdocs==1.5.3              # Documentation generation

# ============================================================================
# DATA PROCESSING
# ============================================================================
pandas==2.1.3              # Data manipulation
numpy==1.26.3              # Numerical computing
pydash==8.0.1              # Utility functions

# ============================================================================
# ASYNC & CONCURRENCY
# ============================================================================
aiohttp==3.9.1             # Async HTTP
asyncio-contextmanager==1.0.0

# ============================================================================
# DEVELOPMENT
# ============================================================================
jupyter==1.0.0             # Notebooks
ipython==8.18.1            # Interactive shell
autopep8==2.0.4            # Code formatter
